add deliveryOption to invoice

+++ database/migrations/2024_04_29_125019_create_glass_widths_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('glass_widths', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('size');
+            $table->decimal('size', 3, 1);
             $table->integer('price');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
